+++ SpacePyTraders/models.py
@@ -1,179 +1,149 @@
+from dataclasses import dataclass, field
+
+@dataclass
 class User ():
-    """The Ship is awesome!"""
-    def __init__(self, data=None, username=None, credits=None, ships=None, loans=None):
-        # Regular Init
-        self.username = username 
-        self.credits = credits
-        self.ships = [Ship(ship) for ship in ships['ships']] if isinstance(ships, dict) else ships
-        self.loans = [Loan(loan) for loan in loan['loans']] if isinstance(loans, dict) else loans
-        # Init with JSON
-        if data is not None and isinstance(data, dict):
-            if 'user' in data:
-                data = data['user']
-            self.username = data['username']
-            self.credits = data['credits']
-            self.ships = [Ship(ship) for ship in data['ships']]
-            self.loans = [Loan(loan) for loan in data['loans']]
-        # Init with User object
-        if data is not None and isinstance(data, User):
-            self.username = data.username
-            self.credits = data.credits
-            self.ships = data.ships
-            self.loans = data.loans
-        
-    def asDict(self):
-        return_dict = {'username': self.username, 'credits': self.credits}
-        return_dict['ships'] = [ship.asDict() for ship in self.ships]
-        return_dict['loans'] = [loan.asDict() for loan in self.loans]
-        return return_dict
-    
-    def __repr__(self):
-        return f"Username: {self.username}, Credits: {self.credits}, No of Ships: {len(self.ships)}, No of Loans: {len(self.loans)}"
-    
-    def __str__(self):
-        return f"Username: {self.username}, Credits: {self.credits}, No of Ships: {len(self.ships)}, No of Loans: {len(self.loans)}"
+    username: str
+    credits: int
+    ships: field(default_factory=list)
+    loans: field(default_factory=list)
 
-
-class Ship ():
-    """The Ships is Awesome!"""
-    def __init__(self, data=None, id=None, manufacturer=None, kind=None, type=None, location=None, x=None, y=None, flightPlanId=None, speed=None, plating=None, weapons=None, maxCargo=None, spaceAvailable=None, cargo=None, **kwargs):
-        """The ship class
+    def __post_init__(self):
+        """Handles creating a list of the respective objects if only the dictionary is given. 
+        This basically means from the get go a you could call `user.ships[0].id` and get the id of the ship back. 
+        Rather than user.ships[0]['id']
         """
-        # Regular init
-        self.id = id
-        self.manufacturer = manufacturer
-        self.kind = kind
-        self.type = type
-        self.location = location
-        self.x = x
-        self.y = y
-        self.flightPlanId = flightPlanId
-        self.speed = speed
-        self.plating = plating
-        self.weapons = weapons
-        self.maxCargo = maxCargo
-        self.spaceAvailable = spaceAvailable
-        self.cargo = cargo
-        # If data is JSON
-        if data is not None and isinstance(data, dict):
-            if 'ship' in data:
-                data = data['ship']
-            if 'location' not in data:
-                self.location = "IN TRANSIT"
-                self.x = None
-                self.y = None
-            if 'flightPlanId' not in data:
-                self.flightPlanId = None
-            self.id = data['id']
-            self.manufacturer = data['manufacturer']
-            self.kind = data['class']
-            self.type = data['type']
-            self.speed = data['speed']
-            self.plating = data['plating']
-            self.weapons = data['weapons']
-            self.maxCargo = data['maxCargo']
-            self.spaceAvailable = data['spaceAvailable']
-            self.cargo = data['cargo']
-        # If data is a Ship object
-        if data is not None and isinstance(data, Ship):
-            self.id = data.id
-            self.manufacturer = data.manufacturer
-            self.kind = data.kind
-            self.type = data.type
-            self.location = data.location
-            self.x = data.x
-            self.y = data.y
-            self.flightPlanId = data.flightPlanId
-            self.speed = data.speed
-            self.plating = data.plating
-            self.weapons = data.weapons
-            self.maxCargo = data.maxCargo
-            self.spaceAvailable = data.spaceAvailable
-            self.cargo = data.cargo
+        if all(isinstance(ship, dict) for ship in self.ships):
+            self.ships = [build_ship(ship) for ship in self.ships]
 
-    def asDict(self):
-        return self.__dict__
-                
-    def __repr__(self):
-        return f"ID: {self.id}, Manufacturer: {self.manufacturer}, Class: {self.kind} " \
-               f"Type: {self.type}, Location: {self.location}, X: {self.x} " \
-               f"Y: {self.y}, Flight Plan ID: {self.flightPlanId}, Speed: {self.speed} " \
-               f"Plating: {self.plating}, Weapons: {self.weapons}, Max Cargo: {self.maxCargo} " \
-               f"Space Available: {self.spaceAvailable}, No of Goods in Cargo: {len(self.cargo)}"
+        if all(isinstance(loan, dict) for loan in self.loans):
+            self.loans = [Loan(**loan) for loan in self.loans]
+
+
+def build_ship(ship_dict):
+    """Handles the creation of a ship class. The ship dict contains a 'class' key which needs to be changed for the class creation.
+    The ship may also be in transit and that needs to be handled accordingly
+
+    Args:
+        ship_dict (dict): the dict version of a ship
+
+    Returns:
+        Ship: A ship object
+    """
+    # Handle if class key is present in dictionary
+    if 'class' in ship_dict:
+        ship_dict['kind'] = ship_dict.pop('class')
+
+    if 'location' not in ship_dict:
+        ship_dict['location'] = "IN_TRANSIT"
+    return Ship(**ship_dict)
+
+@dataclass
+class Ship ():
+    id: str
+    manufacturer: str
+    kind: str
+    type: str
+    location: str
+    speed: int
+    plating: int
+    weapons: int
+    maxCargo: int
+    spaceAvailable: int
+    cargo: field(default_factory=list)
+    flightPlanId: str = None
+    x: int = None
+    y: int = None
+
+    def __post_init__(self):
+        """Handles creating a list of Cargo object in the ship
+        """
+        if all(isinstance(c, dict) for c in self.cargo):
+            self.cargo = [Cargo(**c) for c in self.cargo]
     
-    def __str__(self):
-        return f"ID: {self.id}, Manufacturer: {self.manufacturer}, Class: {self.kind} " \
-               f"Type: {self.type}, Location: {self.location}, X: {self.x} " \
-               f"Y: {self.y}, Flight Plan ID: {self.flightPlanId}, Speed: {self.speed} " \
-               f"Plating: {self.plating}, Weapons: {self.weapons}, Max Cargo: {self.maxCargo} " \
-               f"Space Available: {self.spaceAvailable}, No of Goods in Cargo: {len(self.cargo)}"
+    def calculate_fuel_required(self, dest_dist):
+        calc_fuel = lambda d, p: round((d / 4) + 2 + p)
+        penalties = {
+            "MK-I": 2,
+            "MK-II": 3,
+            "MK-III": 4
+        }
+        penalty = penalties[self.kind]
+        return calc_fuel(dest_dist, penalty)
 
+@dataclass
+class Cargo ():
+    good: str
+    quantity: int
+    totalVolume: int
+
+@dataclass
 class Loan ():
-    """The Ship is awesome!"""
-    def __init__(self, data=None, id=None, due=None, repaymentAmount=None, status=None, type=None):
-        # Regular Init
-        self.id = id 
-        self.due = due
-        self.repaymentAmount = repaymentAmount
-        self.status = status
-        self.type = type
-        # Init with JSON
-        if data is not None and isinstance(data, dict):
-            if 'user' in data:
-                data = data['user']
-            self.id = data['id']
-            self.due = data['due']
-            self.repaymentAmount = data['repaymentAmount']
-            self.status = data['status']
-            self.type = data['type']
-        # Init with User object
-        if data is not None and isinstance(data, User):
-            self.id = data.id
-            self.due = data.due
-            self.repaymentAmount = data.repaymentAmount
-            self.status = data.status
-            self.type = data.type
-        
-    def asDict(self):
-        return {'id': self.id, 'due': self.due, 'repaymentAmount': self.repaymentAmount, 'status': self.status, 'type': self.type}
+    id: str 
+    due: str
+    repaymentAmount: int
+    status: str
+    type: str
     
-    def __repr__(self):
-        return f"id: {self.id}, due: {self.due}, repaymentAmount: {self.repaymentAmount}, status: {self.status}, type: {self.type}"
-    
-    def __str__(self):
-        return f"id: {self.id}, due: {self.due}, repaymentAmount: {self.repaymentAmount}, status: {self.status}, type: {self.type}"
-
+@dataclass
 class Location ():
-    """The Ship is awesome!"""
-    def __init__(self, data=None, symbol=None, type=None, name=None, x=None, y=None):
-        # Regular Init
-        self.symbol = symbol 
-        self.y = y
-        self.name = name
-        self.x = x
-        self.y = y
-        # Init with JSON
-        if data is not None and isinstance(data, dict):
-            if 'user' in data:
-                data = data['user']
-            self.symbol = data['symbol']
-            self.y = data['y']
-            self.name = data['name']
-            self.x = data['x']
-            self.y = data['y']
-        # Init with User object
-        if data is not None and isinstance(data, User):
-            self.symbol = data.symbol
-            self.y = data.y
-            self.name = data.name
-            self.x = data.x
-            self.y = data.y
-        
-    def asDict(self):
-        return {'symbol': self.symbol, 'type': self.type, 'name': self.name, 'x': self.x, 'y': self.y}
-    
-    def __repr__(self):
-        return f"symbol: {self.symbol}, type: {self.type}, name: {self.name}, x: {self.x}, y: {self.y}"
-    
-    def __str__(self):
-        return f"symbol: {self.symbol}, type: {self.type}, name: {self.name}, x: {self.x}, y: {self.y}"
+    symbol: str
+    type: str
+    name: str
+    x: int
+    y: int
+    allowsConstruction: bool
+    structures: field(default_factory=list)
+    messages: list = None
+
+@dataclass
+class Marketplace (Location):
+    marketplace: list = field(default_factory=list)
+
+    def __post_init__(self):
+        """Handles creating a list of Location object in the system
+        """
+        if all(isinstance(good, dict) for good in self.marketplace):
+            self.marketplace = [Good(**good) for good in self.marketplace]
+
+    def get_good(self, symbol):
+        """Returns a Good object for the symbol provided
+
+        Args:
+            symbol (str): Symbol of the good Eg: "FUEL"
+
+        Returns:
+            Good: Good object for the symbol given
+        """
+        return next(good for good in self.marketplace if good.symbol == symbol)
+
+@dataclass
+class Good ():
+    symbol: str
+    volumePerUnit: int
+    pricePerUnit: int
+    spread: int
+    purchasePricePerUnit: int
+    sellPricePerUnit: int
+    quantityAvailable: int
+
+@dataclass
+class System ():
+    locations: field(default_factory=list)
+
+    def __post_init__(self):
+        """Handles creating a list of Location object in the system
+        """
+        if all(isinstance(loc, dict) for loc in self.locations):
+            self.locations = [Location(**loc) for loc in self.locations]
+
+    def get_location(self, symbol):
+        """Returns a Location object for the symbol provided
+
+        Args:
+            symbol (str): Symbol of the location Eg: "OE-PM"
+
+        Returns:
+            Location: Location object for the symbol given
+        """
+        return next(loc for loc in self.locations if loc.symbol == symbol)
+